+++ Demo/Application/Features/PayPalWebCheckoutViewController.swift
@@ -150,6 +150,10 @@ class PayPalWebCheckoutViewController: PaymentButtonBaseViewController {
         request.lineItems = [lineItem]
         request.offerPayLater = payLaterToggle.isOn
         request.intent = newPayPalCheckoutToggle.isOn ? .sale : .authorize
+        request.contactInformation = BTContactInformation(
+            recipientEmail: "some@email.com",
+            recipientPhoneNumber: .init(countryCode: "52", nationalNumber: "123456789")
+        )
 
         payPalClient.tokenize(request) { nonce, error in
             sender.isEnabled = true